refractored code (moved few_shot check and attribute to init in RankLLM)
