+++ src/rank_llm/rerank/pointwise/pointwise_rankllm.py
@@ -42,11 +42,12 @@ class PointwiseRankLLM(RankLLM, ABC):
         filename: str = "",
         batch_size: int = 32,
     ) -> None:
-        super().__init__(model, context_size, prompt_mode)
+        super().__init__(
+            model, context_size, prompt_mode, num_few_shot_examples, few_shot_file
+        )
         self._device = device
         self._filename = filename
         self._batch_size = batch_size
-        self._num_few_shot_examples = num_few_shot_examples
 
     def rerank_batch(
         self,
